tests: Move unit tests to unit module
