+++ patchwork/tests/__init__.py
@@ -9,6 +9,7 @@ import warnings
 TEST_MAIL_DIR = os.path.join(os.path.dirname(__file__), 'mail')
 TEST_PATCH_DIR = os.path.join(os.path.dirname(__file__), 'patches')
 TEST_FUZZ_DIR = os.path.join(os.path.dirname(__file__), 'fuzztests')
+TEST_SERIES_DIR = os.path.join(os.path.dirname(__file__), 'series')
 
 # configure warnings
