+++ patchwork/tests/unit/api/test_event.py
@@ -8,7 +8,7 @@ from django.urls import reverse
 from rest_framework import status
 
 from patchwork.models import Event
-from patchwork.tests.api import utils
+from patchwork.tests.unit.api import utils
 from patchwork.tests.utils import create_check
 from patchwork.tests.utils import create_cover
 from patchwork.tests.utils import create_cover_comment
