+++ patchwork/tests/unit/api/test_patch.py
@@ -12,7 +12,7 @@ from django.urls import reverse
 from rest_framework import status
 
 from patchwork.models import Patch
-from patchwork.tests.api import utils
+from patchwork.tests.unit.api import utils
 from patchwork.tests.utils import create_maintainer
 from patchwork.tests.utils import create_patch
 from patchwork.tests.utils import create_patches
