+++ patchwork/tests/unit/api/test_person.py
@@ -8,7 +8,7 @@ from django.urls import NoReverseMatch
 from django.urls import reverse
 from rest_framework import status
 
-from patchwork.tests.api import utils
+from patchwork.tests.unit.api import utils
 from patchwork.tests.utils import create_maintainer
 from patchwork.tests.utils import create_person
 from patchwork.tests.utils import create_user
