+++ patchwork/tests/unit/test_signals.py
