+++ tox.ini
@@ -31,7 +31,7 @@ passenv =
     DATABASE_NAME
     DJANGO_TEST_PROCESSES
 commands =
-    python {toxinidir}/manage.py test --noinput --parallel -v 2 --timing -- {posargs:patchwork}
+    python {toxinidir}/manage.py test --noinput --parallel -v 2 --timing -- {posargs:patchwork.tests.unit}
 
 [testenv:pep8]
 deps =
@@ -64,8 +64,8 @@ setenv =
     DJANGO_SETTINGS_MODULE = patchwork.settings.dev
 commands =
     coverage erase
-    coverage run --omit=*tox*,patchwork/tests/*.py,manage.py,patchwork/migrations/*.py \
-         --branch {toxinidir}/manage.py test --noinput patchwork
+    coverage run --omit=*tox*,patchwork/tests/unit/*.py,manage.py,patchwork/migrations/*.py \
+         --branch {toxinidir}/manage.py test --noinput patchwork.tests.unit
     coverage report -m
 
 [gh-actions]
